pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wx-gui"
version = "0.1.0"
description = "macOS desktop GUI for controlling WeChat"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PySide6>=6.7",
  "pydantic>=2.7",
  "loguru>=0.7",
  "APScheduler>=3.10",
  "pyautogui>=0.9.54",
  "opencv-python>=4.9",
  "pyobjc-framework-Cocoa>=10.2",
  "pyobjc-framework-ApplicationServices>=10.2",
  "pyobjc-framework-Vision>=11.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
